requirements.txt
adafruit-circuitpython-dht
gpiozero
lgpio
RPi.GPIO
pahomqtt
requests
